Raise concurrency errors properly for glue tables (#1875)
## Which issue does this PR close?
#1868

## What changes are included in this PR?

As referenced in the issue, the optimistic concurrency was not working
correctly as we need to check the glue table version has not incremented
before we make our update, so we can get back the error if there is a
concurrent modification.

This changes the `update_table` to properly handle this case.

Also changed the `load_table` to `load_table_with_version_id`, and
`load_table` now just uses that.

<!--
Provide a summary of the modifications in this PR. List the main changes
such as new features, bug fixes, refactoring, or any other updates.
-->

## Are these changes tested?

I have tested them using my specific setup with concurrent writers to
iceberg tables in aws glue, but not generically. Writing a test case
doesn't seem feasible as one needs a glue table.
<!--
Specify what test covers (unit test, integration test, etc.).

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

---------

Co-authored-by: Jem Bishop <jem@lo.tech>
# Conflicts:
#	crates/catalog/glue/src/catalog.rs